Removed guessing of best optimizing CFLAGS.
With old versions of GCC "-O1 -fomit-frame-pointer" produced the best
code for x86 systems, but with GCC 4.x this is no longer the case.

With GCC 4.1.0, building with "-O1 -fomit-frame-pointer" makes TRE
about 15% slower than when built with "-O2 -g" (the default).
Removing the "-g" has a negligible positive effect, so it's better to
leave it in so the result is actually debuggable.

darcs-hash:20060722115947-ced27-23242246b62f51ef08efd91e501e7a9756a88411.gz